krsi: Add krsi_is_procfs_file_op helper
This helper can be attached to the "file_open" KRSI hook and allows the
caller to check if a process opened the given file in procfs. is_self_op
checks if the current process is accessing the file.

The helper returns 0 in case the open operation is not for the file in
question and the PID of the file i.e. /proc/<pid> on success.