prototype: eBPF helper to dump arg pages of a process.
bpf_lsm_output_argv dumps the contents of the args buffer into the perf
events buffer.

There are few things that undecided yet about this:

Should this be a part of the bpf_lsm_event_output helper which would
take a flag as an input for standard events provided by KRSI like this
one. There are special versions of bpf_perf_event_outputf for skb and
xdp, but they are not multiplexed with flags.

Once, we sleepable, eBPF this helper can be greatly simplified.

Change-Id: I810cddb0a8baf7fd7457484c8e2adcda4b3204cd