prototype: eBPF helper to check acceses to /proc/<pid>/*
This helper can be attached to the "file_open" LSM Hook and allows the
caller to check if a process opened the given file in procfs. is_self_op
checks if the current process is accessing the file.

The helper returns 0 in case the open operation is not for the file in
question and the PID of the file i.e. /proc/<pid> on success.

Change-Id: I17b10d9c5db2e6715dcf4a899d7c884a83ab34cd