prototype: Add pre/post callbacks for bprm_check_security
The pre-callback allocates a buffer and copies the arg pages and
adds a pointer to the arg_pages in the security blob. This
makes the arg_pages available to BPF helpers which, without this,
cannot have access to this as they run in atomic context and
cannot call get_user_pages_remote.

The post callback frees the allocated pages.

This becomes simpler in the age of sleepable eBPF helpers.`

Change-Id: I8fecba4b4a718edbfd201d25cb139a42bae38322