bpf: lsm: Implement attach, detach and execution
JITed BPF programs are used used by the BPF LSM as dynamically
allocated security hooks. arch_bpf_prepare_trampoline handles the
conversion of the signature of the hook to the BPF context and invokes
the BPF program.

The following permissions are required to attach a program to a hook:

- CAP_SYS_ADMIN to load the program
- CAP_MAC_ADMIN to attach it (i.e. to update the security policy)

When the program is loaded (BPF_PROG_LOAD), the verifier receives the
index of the member in the security_hook_heads struct in the
prog->aux-attach_btf_id and uses the BTF API provided by the LSM to:

- Populate the name of the hook in prog->aux->attach_func_name and
  the prototype in prog->aux->attach_func_proto.
- Verify if the offset is valid for a type struct hlist_head.
- The program is verified for accesses based on the attach_func_proto
  similar to raw_tp BPF programs.

When an attachment (BPF_PROG_ATTACH) is requested:

- The information required to set-up of a callback is populated in the
  struct bpf_lsm_list.
- A new trampoline and a bpf_lsm_hook is allocated and the address of
  the trampoline is set to the be the address of the hook.
- The attachment returns an anonymous O_CLOEXEC fd which detaches the
  program on close.

Change-Id: I3a4b7bf87f653437b98a935167dfb07b36a34004