security/loadpin: Expose 'loadpin.enforce' as 'loadpin.enabled'
COS depends on loadpin.enabled heavily on the following:
  1.) COS customizer/Preloader
  2.) Module signing
  3.) Test containers
  4.) Documents
To avoid regression and minimize kernel merging conflicts, this
patch provides the best way to handle this situation for now.
This patch can be removed after 4.19.x kernel is deprecated or
we update places mentioned above.

BUG=b/155888357
TEST=tryjob
RELEASE_NOTE=None
SOURCE=COS

Change-Id: I8e4f1f3caf0a2abb7ed31b2d8744056114271258
Signed-off-by: Roy Yang <royyang@google.com>
Reviewed-on: https://cos-review.googlesource.com/c/third_party/kernel/+/1460
Reviewed-by: Vaibhav Rustagi <vaibhavrustagi@google.com>