sched/futex: add wake_up_swap, use with GFUTEX_SWAP
As described in the previous patch in this patchset, it is
often beneficial to wake a task and run it on the same CPU
where the current (=going to sleep) task it running.

This patch adds wake flag WF_CURRENT_CPU which is then used
by wake_up_swap to hit that the wakee should be migrated
to the current CPU. select_task_rq_fair then takes the hint.

With this patch the benchmark from the next patch in this
patchset shows the following:

$ ./futex_swap -i 1000000

------- running SWAP_WAKE_WAIT -----------

completed 1000000 swap and back iterations in 14506367240 ns: 7253 ns
per swap
PASS

------- running SWAP_SWAP -----------

completed 1000000 swap and back iterations in 2035672225 ns: 1017 ns per
swap
PASS

BUG=b/153575851
TEST=compile
SOURCE=kernel/kubi:core/5.4/extensions/sched/fibers
RELEASE_NOTE=Port sched/futex enhancement to COS kernel

Signed-off-by: Peter Oskolkov <posk@google.com>
Change-Id: Idfcb83b00193260281dfed6b6dea4594e49b88fa
Reviewed-on: https://cos-review.googlesource.com/c/third_party/kernel/+/8901
Reviewed-by: Vaibhav Rustagi <vaibhavrustagi@google.com>
Tested-by: Cusky Presubmit Bot <presubmit@cos-infra-prod.iam.gserviceaccount.com>