security: replace indirect calls with static calls
LSMs have high overhead due to indirect function calls through
retpolines. This RPC proposes to replace these with static calls [1]
instead.

Indirect function calls have a high overhead due to retpoline
mitigations for spectre_v2 .

This overhead is especially significant for the "bpf" LSM which supports
the implementation of LSM hoaoks with eBPF programs (security/bpf)[2]. In
order to facilitate this, the "bpf" LSM provides a default callback for
all LSM hooks. When enabled, the "bpf", LSM incurs an unnecessary /
avoidable indirect call. This is especially worse in OS hot paths in the
networking stack.

The performance impact on a simple syscall eventfd_write (which triggers
the file_permission hook) was measured with and without "bpf" LSM
enabled. Activating the LSM resulted in an overhead of 4% [3].

This overhead prevents the adoption of bpf LSM on performance critical
systems, and also, in general, slows down all LSMs.

Currently, the LSM hook callbacks are stored in a linked list and
dispatched as indirect calls. Using static calls can remove this overhead
by replacing all indirect calls with direct calls.

During the discussion of the "bpf" LSM patch-set it was proposed to special
case BPF LSM to avoid the overhead by using static keys. This was however
not accepted and it was decided to [4]:

- Not special-case the "bpf" LSM.
- Implement a general solution benefitting the whole LSM framework.

The patch replaces the linked list implementation and introduces
security slots. The security slots have the following information:

1. A static call which is updated to the relevant LSM callback on
   system initialization once the order of LSMs is known.
2. A static key which indicates if the slot is active (i.e. has an
   LSM callback enabled.
3. Information regarding the default state of the slot. LSMs can
   register callbacks in a default_disabled state too.

The patch also adds a macro SECURITY_FOREACH_STATIC_SLOT(MACRO, x, y) ->

	MACRO(0, x, y)
	MACRO(1, x, y)
	MACRO(2, x, y)
	MACRO(3, x, y)
	MACRO(4, x, y)

This is used in conjunction with LSM_HOOK definitions in
linux/lsm_hook_defs.h to execute a macro for each static slot of each LSM
hook.

The number of available slots for each LSM hook is currently fixed at
11 (the number of LSMs in the kernel).

[1]: https://lwn.net/ml/linux-kernel/20200710133831.943894387@infradead.org/
[2]: https://lwn.net/Articles/798157/
[3] measurements: https://gist.githubusercontent.com/PaulRenauld/fe3ee7b51121556e03c181432c8b3dd5/raw/62437b1416829ca0e8a0ed9101530bc90fd42d69/lsm-performance.png
protocol: https://gist.github.com/PaulRenauld/fe3ee7b51121556e03c181432c8b3dd5#file-measurement-protocol-md
[4]: https://lwn.net/Articles/813261/
[5]: git://git.kernel.org/pub/scm/linux/kernel/git/peterz/queue.git x86/static_call
[6]: https://lwn.net/ml/linux-kernel/20200710133831.943894387@infradead.org/#t

Cc: Alexei Starovoitov <ast@kernel.org>
Cc: Daniel Borkmann <daniel@iogearbox.net>
Cc: James Morris <jmorris@namei.org>
Cc: pjt@google.com
Cc: jannh@google.com
Cc: peterz@infradead.org
Cc: rafael.j.wysocki@intel.com
Cc: keescook@chromium.org
Cc: thgarnie@chromium.org
Cc: kpsingh@google.com

Co-developed-by: Paul Renauld <renauld@google.com>
Co-developed-by: KP Singh <kpsingh@google.com>
Co-developed-by: Brendan Jackman <jackmanb@google.com>
Change-Id: I81955b3ab8e495bc08193b62420f79b25dbb4493