bpf: Implement support for BPF program signatures
Signatures are used to verify the integrity of BPF programs
which help ensure that BPF programs were built from a trusted
source.

Unlike kernel modules, the signature verification for BPF programs
aims at providing a flexible interface. In order to provide a flexible
interface, a void pointer and size is introduced to the UAPI for
BPF_PROG_LOAD.

BPF program signing requires a stable instruction buffer and, CO-RE
programs perform relocations after build time which modify the
instruction buffer and invalidate the signature. To address this,
light skeletons were introduced which generates a loader program
capturing the various syscall operations including the original
instruction buffer.

Since these syscalls are performed via helper (bpf_sys_bpf)
which are available to loader progam. To address the pointer
namespaces for these two mode (bpf syscall from user and kernel)
bpfptr_t support was added.

In order to provide metadata to LSM to indicate whether a particular
bpf_prog_load was generated by userspace or kernel, this information is
added to bpf_prog_aux as is_kernel.

The signatuere and instruction data are stored as dynptrs in the
prog->aux struct as the kfunc for verifying signature uses dynptrs and
it's not possible to construct dynptrs from kernel void * pointers in
BPF programs.